// src/app/Pages/Models/smsservice.ts
export class smsservice {
    ID: number;
    CLIENT_ID: number;
    SERVICE_PROVIDER: any;
    API_URL: any;
    API_KEY: any;
    AUTHENTICATION_TYPE: any;
    USERNAME: any;
    PASSWORD: any;
    SENDER_ID: any;
    DEFAULT_COUNTRY_CODE: any;
    RETRY_ATTEMPTS = 0;
    TIMEOUT_SECONDS: any;
    IS_ACTIVE: boolean = true;
    CREATED_AT: any;
    UPDATED_AT: any;
    NOTES: any;
    SENDER_EMAIL: any;
    SENDER_NAME: any;
}